Backend/requirements.txt
fastapi==0.115.12
strawberry-graphql==0.266.0
uvicorn==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0
redis==5.0.1
fastapi-cache2==0.2.1
boto3==1.28.38
python-multipart==0.0.6